cleaned up folder structure  and refactor oneWayBF.
